Test CDI Managed Bean bindings
Test use of an ibm-managed-bean-bnd.xml file, and the corresponding
server.xml config, with CDI managed beans.

Add Xalan serializer to dependencies. Another test in this bucket
requires HtmlUnit which requires Xalan. The serializer dependency is
needed when we use the ServerConfiguration model, and the server.xml
includes elements which aren't part of our model, such as
<managed-bean-bnd> and its decendents.